requirements.txt
python-telegram-bot
beautifulsoup4
lxml
requests
